Decouple rm_paths and rm_parents
Technically, we don't need to know the lot_name ahead of time when we only
want to remove paths from the database, because the SQL table is set to
prohibit duplicate paths. As such, I decoupled these two APIs so that
we can call the rm_paths API without specifying the lot name. This will
be useful in the case of removing spool dirs/files in HTCondor, because
the function that handles the actual removal has no way of telling which
lot the dirs/files belong to without asking lotman.